change schema

+++ frontend/src/components/ContactBox.js
@@ -6,15 +6,18 @@ export default function ContactBox (props) {
     const handleClick = async () => {
         const response = await fetch('/api/contacts/' + props.contact._id, {method: "DELETE"})
         const contact = await response.json()
-        console.log(contact)
-        dispatch({type: "DELETE", payload: contact._id})
+
+        if(response.ok) {
+            dispatch({type: "DELETE", payload: contact._id})
+        }
     }
 
     return (
         <div className="contact-box">
             <h4>{props.contact.name}</h4>
-            <p>{props.contact.phone}</p>
-            <p>{props.contact.address}</p>
+            <p>{props.contact.phone && `Phone: ${props.contact.phone}`}</p>
+            <p>{props.contact.email && `Email: ${props.contact.email}`}</p>
+            <p>{props.contact.address && `Address: ${props.contact.address}`}</p>
             <span onClick={handleClick}>Delete</span>
         </div>
     )
